refactor EditProfile Component
